Moved textarea, checkbox, toggle and radio out of input folder
