+++ packages/react-components/src/button/stories/button.stories.mdx
@@ -1,6 +1,6 @@
 import { AddIcon, SignoutIcon } from "@react-components/icons";
 import { Button, PureButton } from "@react-components/button";
-import { HocWarning, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { Label, Tag } from "@react-components/label";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
@@ -21,7 +21,7 @@ A Button indicates a possible user action.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="input" imports={<Imports />}/>
+<InstallReactComponents imports={<Imports />}/>
 
 ## Props
