+++ packages/react-components/src/checkbox/index.js
@@ -1,0 +1,1 @@
+export * from "./src";
