+++ packages/react-components/src/checkbox/src/index.js
@@ -1,0 +1,1 @@
+export * from "./checkbox";
