+++ packages/react-components/src/checkbox/stories/checkbox.stories.mdx
@@ -1,6 +1,6 @@
-import { Checkbox, PureCheckbox } from "@react-components/input";
+import { Checkbox, PureCheckbox } from "@react-components/checkbox";
 import { CommunicationIcon, GroupIcon, InfoIcon } from "@react-components/icons";
-import { HocWarning, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { Label } from "@react-components/label";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
@@ -19,7 +19,7 @@ A checkbox allows a user to select a value from a small set of options, often bi
 
 ## Installation
 
-<InstallationReactComponents componentFolder="input" imports={<Imports />}/>
+<InstallReactComponents imports={<Imports />}/>
 
 ## Props
