+++ packages/react-components/src/date-picker/stories/date-range-picker/date-range-picker.stories.mdx
@@ -12,7 +12,7 @@ import {
 import { ControlledDateRangePicker } from "./components";
 import { DEFAULT_PRESETS_OPTIONS } from "../data";
 import { DEFAULT_SIZE } from "@react-components/date-picker/src/sizes";
-import { FilePreview, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { FilePreview, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { LARGE, MEDIUM, SMALL } from "@react-components/shared";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { boolean, select, text, withKnobs } from "@storybook/addon-knobs";
@@ -36,7 +36,7 @@ A date range picker allow a user to select a date range.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="date-picker" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
