+++ packages/react-components/src/date-picker/stories/inline-single-date-picker/inline-single-date-picker.stories.mdx
@@ -7,7 +7,7 @@ import {
     TOP_LEFT,
     TOP_RIGHT
 } from "@react-components/date-picker";
-import { InstallationReactComponents, Props, ReactComponentsLayout, WarningMessage } from "@blocks";
+import { InstallReactComponents, Props, ReactComponentsLayout, WarningMessage } from "@blocks";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { boolean, number, select, text, withKnobs } from "@storybook/addon-knobs";
 import { createComponentSection, paramsBuilder } from "@utils";
@@ -29,7 +29,7 @@ An inlined single date picker can appear inline with other content.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="date-picker" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
