+++ packages/react-components/src/date-picker/stories/single-date-picker/single-date-picker.stories.mdx
@@ -10,7 +10,7 @@ import {
 } from "@react-components/date-picker";
 import { ControlledSingleDatePicker } from "./components";
 import { DEFAULT_SIZE } from "@react-components/date-picker/src/sizes";
-import { FilePreview, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { FilePreview, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { LARGE, MEDIUM, SMALL } from "@react-components/shared";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { boolean, number, select, text, withKnobs } from "@storybook/addon-knobs";
@@ -34,7 +34,7 @@ A single date picker allow a user to select a date.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="date-picker" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
