+++ packages/react-components/src/icons/stories/icon.stories.mdx
@@ -1,6 +1,6 @@
 import { AzureIcon32, FilterIcon24, FilterIcon32 } from "./assets";
 import { BIG, HUGE, LARGE, MASSIVE, MEDIUM, SMALL, TINY } from "@react-components/shared";
-import { CheckeredBackground, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { CheckeredBackground, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { DEFAULT_SIZE, Icon, MultiVariantIcon, PureIcon, PureMultiVariantIcon } from "@react-components/icons";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
@@ -23,7 +23,7 @@ If you are looking for Orbit UI official icons, please have a look at the [icons
 
 ## Installation
 
-<InstallationReactComponents componentFolder="icons" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
