+++ packages/react-components/src/index.js
@@ -3,6 +3,10 @@ export * from "./date-picker";
 export * from "./dropdown";
 export * from "./icons";
 export * from "./input";
+export * from "./textarea";
+export * from "./checkbox";
+export * from "./toggle";
+export * from "./radio";
 export * from "./label";
 export * from "./multi-select";
 export * from "./popup";
