+++ packages/react-components/src/input/src/index.js
@@ -1,5 +1,1 @@
 export * from "./input";
-export * from "./textarea";
-export * from "./checkbox";
-export * from "./radio";
-export * from "./toggle";
