+++ packages/react-components/src/input/stories/input/input.stories.mdx
@@ -1,6 +1,6 @@
 import { Button } from "@react-components/button";
 import { CloseIcon, MagnifierIcon } from "@react-components/icons";
-import { HocWarning, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { Input, PureInput } from "@react-components/input";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
@@ -18,7 +18,7 @@ import Imports from "./imports.mdx";
 
 ## Installation
 
-<InstallationReactComponents componentFolder="input" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
