+++ packages/react-components/src/label/stories/label/label.stories.mdx
@@ -1,6 +1,6 @@
 import { Button } from "@react-components/button";
 import { CloseIcon, CommunicationIcon, MagnifierIcon } from "@react-components/icons";
-import { HocWarning, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { Label, PureLabel, Tag } from "@react-components/label";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
@@ -21,7 +21,7 @@ A label displays content classification.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="label" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
