+++ packages/react-components/src/label/stories/tag/tag.stories.mdx
@@ -1,4 +1,4 @@
-import { HocWarning, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { PureTag, Tag } from "@react-components/label";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
@@ -18,7 +18,7 @@ A tag is a small colored dot displayed before the text.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="label" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
