+++ packages/react-components/src/multi-select/stories/multi-select.stories.mdx
@@ -1,7 +1,7 @@
 import { ControlledMultiSelect } from "./components";
 import { DEFAULT_ITEMS, DEFAULT_ITEMS_WITH_CATEGORIES, GROUP_NAME_CHANGED_VALUE, GROUP_RESTORED_VALUE } from "./data";
 import { DEFAULT_SIZE, MultiSelect } from "@react-components/multi-select";
-import { FilePreview, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { FilePreview, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { LARGE, MEDIUM, SMALL } from "@react-components/shared";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { boolean, select, text, withKnobs } from "@storybook/addon-knobs";
@@ -24,7 +24,7 @@ A multi-select allow a user to select multiple values from a fixed set of data.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="multi-select" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
