+++ packages/react-components/src/popup/stories/popup/popup.stories.mdx
@@ -1,6 +1,6 @@
 import { BOTTOM_CENTER, BOTTOM_LEFT, BOTTOM_RIGHT, Popup, TOP_CENTER, TOP_LEFT, TOP_RIGHT } from "@react-components/popup";
 import { ControlledRedBox, RedBoxPopup } from "./components";
-import { InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { array, select, withKnobs } from "@storybook/addon-knobs";
 import { createComponentSection, paramsBuilder } from "@utils";
@@ -17,7 +17,7 @@ import Imports from "./imports.mdx";
 
 ## Installation
 
-<InstallationReactComponents componentFolder="popup" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
