+++ packages/react-components/src/popup/stories/tooltip/tooltip.stories.mdx
@@ -1,6 +1,6 @@
 import { AddIcon } from "@react-components/icons";
 import { Button } from "@react-components/button";
-import { HocWarning, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { Label } from "@react-components/label";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { PureTooltip, Tooltip } from "@react-components/popup";
@@ -18,7 +18,7 @@ import Imports from "./imports.mdx";
 
 ## Installation
 
-<InstallationReactComponents componentFolder="popup" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
