+++ packages/react-components/src/radio/src/radio.jsx
@@ -1,4 +1,4 @@
-import { Checkbox } from "../../input";
+import { Checkbox } from "../../checkbox";
 import { Ref } from "semantic-ui-react";
 import { arrayOf, bool, element, func, number, object, oneOf, oneOfType, string } from "prop-types";
 import { forwardRef } from "react";
@@ -11,7 +11,7 @@ const DEFAULT_SIZE = "medium";
 // Duplicated here until https://github.com/reactjs/react-docgen/pull/352 is merged. Otherwise the preset will not render properly in the docs.
 const CHECKBOX_PROP_TYPES = {
     /**
-     * Whether or not the toggle should autofocus on render.
+     * Whether or not the radio should autofocus on render.
      */
     autofocus: bool,
     /**
@@ -19,15 +19,15 @@ const CHECKBOX_PROP_TYPES = {
      */
     autofocusDelay: number,
     /**
-     * The text associated to the toggle.
+     * The text associated to the radio.
      */
     text: string,
     /**
-     * A toggle can display icons.
+     * A checkbox can display icons.
      */
     icons: oneOfType([element, arrayOf(element)]),
     /**
-     * A label displayed after the toggle text.
+     * A label displayed after the radio text.
      */
     label: oneOfType([element, object]),
     /**
@@ -64,36 +64,37 @@ const CHECKBOX_DEFAULT_PROPS = {
     unsupportedPropsComponentName: "@orbit-ui/react-components/checkbox"
 };
 
-const UNSUPPORTED_PROPS = ["as", "defaultIndeterminate", "indeterminate", "slider", "radio", "type"];
+const UNSUPPORTED_PROPS = ["as", "defaultIndeterminate", "indeterminate", "slider", "toggle", "type"];
 
-export function PureToggle(props) {
+export function PureRadio(props) {
     const { forwardedRef, ...rest } = props;
 
     const renderWithRef = () => {
         return (
             <Ref innerRef={forwardedRef}>
-                {renderToggle()}
+                {renderRadio()}
             </Ref>
         );
     };
 
-    const renderToggle = () => {
+    const renderRadio = () => {
         return (
             <Checkbox
-                toggle
+                radio
                 unsupportedProps={UNSUPPORTED_PROPS}
-                unsupportedPropsComponentName="@orbit-ui/react-components/toggle"
+                unsupportedPropsComponentName="@orbit-ui/react-components/radio"
                 {...rest}
             />
         );
     };
 
-    return isNil(forwardedRef) ? renderToggle() : renderWithRef();
+    return isNil(forwardedRef) ? renderRadio() : renderWithRef();
 }
 
-PureToggle.propTypes = CHECKBOX_PROP_TYPES;
-PureToggle.defaultProps = CHECKBOX_DEFAULT_PROPS;
+PureRadio.propTypes = CHECKBOX_PROP_TYPES;
+PureRadio.defaultProps = CHECKBOX_DEFAULT_PROPS;
 
-export const Toggle = forwardRef((props, ref) => (
-    <PureToggle { ...props } forwardedRef={ref} />
+export const Radio = forwardRef((props, ref) => (
+    <PureRadio { ...props } forwardedRef={ref} />
 ));
+
