+++ packages/react-components/src/radio/stories/radio.stories.mdx
@@ -1,7 +1,7 @@
 import { ControlledRadioGroup } from "./components";
-import { FilePreview, HocWarning, InstallationReactComponents, Props, ReactComponentsLayout, WarningMessage } from "@blocks";
+import { FilePreview, HocWarning, InstallReactComponents, Props, ReactComponentsLayout, WarningMessage } from "@blocks";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
-import { PureRadio, Radio } from "@react-components/input";
+import { PureRadio, Radio } from "@react-components/radio";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
 import Imports from "./imports.mdx";
 
@@ -18,7 +18,7 @@ Radio buttons allow a user to select one option from a set.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="input" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
