+++ packages/react-components/src/radio/tests/jest/radio.test.jsx
@@ -1,4 +1,4 @@
-import { Radio } from "@react-components/input";
+import { Radio } from "@react-components/radio";
 import { createRef } from "react";
 import { render, wait } from "@testing-library/react";
