+++ packages/react-components/src/search-input/stories/remote-search-input/remote-search-input.stories.mdx
@@ -1,6 +1,6 @@
 import { ControlledRemoteSearchInput } from "./components";
 import { DEFAULT_SIZE, RemoteSearchInput, searchInputResult, useDefaultResultsFetcher } from "@react-components/search-input";
-import { FilePreview, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { FilePreview, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { LARGE, MEDIUM, SMALL } from "@react-components/shared";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { boolean, number, select, text, withKnobs } from "@storybook/addon-knobs";
@@ -24,7 +24,7 @@ A remote search input allows a user to query for results from a remote data sour
 
 ## Installation
 
-<InstallationReactComponents componentFolder="search-input" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
