+++ packages/react-components/src/search-input/stories/search-input/search-input.stories.mdx
@@ -1,7 +1,7 @@
 import { ControlledSearchInput } from "./components";
 import { DEFAULT_RESULTS, LAURIE_VALUE } from "../data";
 import { DEFAULT_SIZE, SearchInput, SearchInputController } from "@react-components/search-input";
-import { FilePreview, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { FilePreview, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { LARGE, MEDIUM, SMALL } from "@react-components/shared";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { boolean, number, select, text, withKnobs } from "@storybook/addon-knobs";
@@ -24,7 +24,7 @@ A search input allows a user to query for results from a fixed set of data.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="search-input" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
