+++ packages/react-components/src/select/stories/inline.stories.mdx
@@ -1,4 +1,4 @@
-import { HocWarning, InstallationReactComponents, Props, ReactComponentsLayout, WarningMessage } from "@blocks";
+import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout, WarningMessage } from "@blocks";
 import { MagnifierIcon } from "@react-components/icons";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { PureSelect, Select, selectItem } from "@react-components/select";
@@ -19,7 +19,7 @@ An inlined select can appear inline with other content.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="select" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
