+++ packages/react-components/src/select/stories/items.stories.mdx
@@ -1,5 +1,5 @@
 import { InfoIcon, WarningIcon } from "@react-components/icons";
-import { InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { PureSelect, Select, selectItem } from "@react-components/select";
 import { createComponentSection, paramsBuilder } from "@utils";
@@ -20,7 +20,7 @@ An item of a select component.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="select" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
