+++ packages/react-components/src/select/stories/multiples.stories.mdx
@@ -1,4 +1,4 @@
-import { HocWarning, InstallationReactComponents, Props, ReactComponentsLayout, WarningMessage } from "@blocks";
+import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout, WarningMessage } from "@blocks";
 import { MagnifierIcon } from "@react-components/icons";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { PureSelect, Select, selectItem } from "@react-components/select";
@@ -20,7 +20,7 @@ A select can allow a user to make multiple choices.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="select" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
