+++ packages/react-components/src/select/stories/standard.stories.mdx
@@ -1,4 +1,4 @@
-import { HocWarning, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { MagnifierIcon } from "@react-components/icons";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { PureSelect, Select, selectItem } from "@react-components/select";
@@ -20,7 +20,7 @@ A select can be used to select between choices.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="select" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ### Props
