+++ packages/react-components/src/textarea/index.js
@@ -1,0 +1,1 @@
+export * from "./src";
