+++ packages/react-components/src/textarea/src/index.js
@@ -1,0 +1,1 @@
+export * from "./textarea";
