+++ packages/react-components/src/textarea/stories/textarea.stories.mdx
@@ -1,6 +1,6 @@
-import { HocWarning, InstallationReactComponents, Props, ReactComponentsLayout } from "@blocks";
+import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
-import { PureTextArea, TextArea } from "@react-components/input";
+import { PureTextArea, TextArea } from "@react-components/textarea";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
 import Imports from "./imports.mdx";
 
@@ -18,7 +18,7 @@ A textarea can be used to allow for extended user input.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="input" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
