+++ packages/react-components/src/toggle/stories/toggle.stories.mdx
@@ -1,6 +1,6 @@
-import { HocWarning, InstallationReactComponents, Props, ReactComponentsLayout, WarningMessage } from "@blocks";
+import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout, WarningMessage } from "@blocks";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
-import { PureToggle, Toggle } from "@react-components/input";
+import { PureToggle, Toggle } from "@react-components/toggle";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
 import Imports from "./imports.mdx";
 
@@ -17,7 +17,7 @@ A toggle offers an on or off choice to a user.
 
 ## Installation
 
-<InstallationReactComponents componentFolder="input" imports={<Imports />} />
+<InstallReactComponents imports={<Imports />} />
 
 ## Props
